Some refactoring of JS objects to use new JsValue API, fixes crashes in setTimeout() as a side-effect

+++ include/AdblockPlus/WebRequest.h
@@ -10,7 +10,7 @@ namespace AdblockPlus
 
   struct ServerResponse
   {
-    unsigned int status;
+    int64_t status;
     HeaderList responseHeaders;
     int responseStatus;
     std::string responseText;
